Collisonクラスを少し改良（円形衝突判定）ゲームループを管理するmaintaskクラスの追加

+++ src/src/CannonArtilally.cpp
@@ -38,7 +38,7 @@ void CannonArtilally::DevideAmmo_rapid(int rate)
 {
 	ammo_rapid %= rate;
 }
-void CannonArtilally::Instantiate(double m_Angle,int m_x,int m_y)
+void CannonArtilally::Instantiate(double m_Angle,double m_x,double m_y)
 {
 	Angle = m_Angle;
 	position.x = m_x;
